+++ package.json
@@ -26,10 +26,10 @@
     "marked": "^0.6.0",
     "moment": "^2.22.2",
     "query-string": "^6.2.0",
-    "react": "^16.4.2",
+    "react": "^16.8.1",
     "react-apollo": "^2.2.4",
     "react-datetime-picker": "^2.1.1",
-    "react-dom": "^16.4.2",
+    "react-dom": "^16.8.1",
     "react-emotion": "^9.2.6",
     "react-router": "^4.3.1",
     "react-router-dom": "^4.3.1",
