+++ src/layout/MobileMenu.js
@@ -1,0 +1,5 @@
+import React from 'react'
+
+function MobileMenu() {}
+
+export default MobileMenu
